refactor(routes): move route constants to dedicated routes file
