+++ src/routes/routes.js
@@ -1,0 +1,21 @@
+export const NAVIVATION_LINKS = [
+  { title: "Home", url: "/" },
+  { title: "Movies", url: "/movies" },
+  { title: "Favourite", url: "/favourite" },
+];
+
+export const GENRES = [
+  { title: "Action", url: "/genre/action" },
+  { title: "Comedy", url: "/genre/comedy" },
+  { title: "Drama", url: "/genre/drama" },
+  { title: "Sci-Fi", url: "/genre/sci-fi" },
+  { title: "Horror", url: "/genre/horror" },
+];
+
+export const LEGAL = [
+  { title: "Terms of Service", url: "/terms" },
+  { title: "Privacy Policy", url: "/privacy" },
+  { title: "Cookie Policy", url: "/cookies" },
+  { title: "FAQ", url: "/faq" },
+  { title: "Help Center", url: "/help" },
+];
